requirements.txt
# Core ML/DL Framework - CUDA 12.1 compatible
--index-url https://download.pytorch.org/whl/cu121
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

# Transformers & Model Utilities
transformers==4.45.0
accelerate==0.30.1
safetensors==0.4.3
tokenizers==0.20.0
huggingface-hub==0.25.0

# Data Processing & Scientific Computing
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

# Data Storage & Serialization
pyarrow==15.0.2
duckdb==0.10.3
sqlalchemy==2.0.30
alembic==1.13.1

# Schema & Config Management
pydantic==2.7.3
pydantic-core==2.18.4
pyyaml==6.0.1

# Graph & Network Analysis
networkx==3.2.1

# CLI & UI
rich==13.7.1
tqdm==4.66.4
click==8.1.7

# Testing & Development
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
ruff==0.4.7

# Type Checking
mypy==1.10.0
types-PyYAML==6.0.12.20240311

# Additional utilities
python-dateutil==2.9.0.post0
typing-extensions==4.11.0
filelock==3.14.0
fsspec==2024.3.1
packaging==24.0
requests==2.32.3
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7

# Optional: Jupyter/Notebook support (uncomment if needed)
# jupyter==1.0.0
# ipython==8.24.0
# notebook==7.1.3

# Optional: Visualization (uncomment if needed)
# matplotlib==3.8.4
# seaborn==0.13.2
# plotly==5.21.0

# Viewer bridge (WebSocket server)
websockets==12.0
google-generativeai==0.5.4
